added confimation / group no

+++ admin/manage_faculty.php
@@ -284,7 +284,8 @@ if(isset($_POST['save'])){
                                     <td>
                                         <div class="actions">
                                             <a class="action-edit" href="#"><i class="ri-edit-line"></i></a>
-                                             <a class="action-delete" href="delete_faculty.php?id=<?php echo $value['id'] ?>" ><i class="ri-delete-bin-line"></i></a>
+                                             <!--<a class="action-delete" href="delete_faculty.php?id=<?php echo $value['id'] ?>" ><i class="ri-delete-bin-line"></i></a>-->
+                                             <a class="action-delete" href="delete_faculty.php?id=<?php echo $value['id'] ?>" onclick="return confirm('Are you sure to delete?')"><i class="ri-delete-bin-line"></i></a>
                                         </div>
                                     </td>
                                 <?php
@@ -382,6 +383,14 @@ if(isset($_POST['save'])){
     <script src="https://cdn.datatables.net/1.13.1/js/dataTables.bootstrap5.min.js"></script>
     <!--responsive-->
     <script src="https://cdn.datatables.net/responsive/2.4.0/js/dataTables.responsive.min.js"></script>
+    <script>
+    function confirmation(){
+      var result = confirm("Are you sure to delete?");
+      if(result){
+        console.log("Deleted")
+      }
+    }
+  </script>
     <!-- end: JS -->
   </body>
 </html>
